Moved fields to separate files for easier tracking.
