+++ api/model/form/country_field.go
@@ -1,0 +1,14 @@
+package form
+
+import "strings"
+
+// CountryField stores a country location
+type CountryField string
+
+func (f CountryField) Valid() (bool, error) {
+
+	if strings.TrimSpace(string(f)) == "" {
+		return false, ErrInvalidLocation{"Country is required", nil}
+	}
+	return true, nil
+}
