+++ api/model/form/country_field_test.go
@@ -1,0 +1,19 @@
+package form
+
+import "testing"
+
+func TestCountryField(t *testing.T) {
+	tests := []struct {
+		Field    CountryField
+		Expected bool
+	}{
+		{"", false},
+		{"United States", true},
+	}
+
+	for _, test := range tests {
+		if ok, _ := test.Field.Valid(); ok != test.Expected {
+			t.Errorf("Expected [%v] to be [%v]\n", ok, test.Expected)
+		}
+	}
+}
