+++ api/model/form/email_field.go
@@ -1,0 +1,13 @@
+package form
+
+import "net/mail"
+
+// EmailField contains a properly formatted email
+type EmailField string
+
+func (e EmailField) Valid() (bool, error) {
+	if _, err := mail.ParseAddress(string(e)); err != nil {
+		return false, ErrFieldInvalid{"Invalid email"}
+	}
+	return true, nil
+}
