+++ api/model/form/employment_status_field.go
@@ -1,0 +1,9 @@
+package form
+
+// EmploymentStatusField contains a person employement status.
+// {Full-time, Part-time}
+type EmploymentStatusField string
+
+func (e EmploymentStatusField) Valid() (bool, error) {
+	return true, nil
+}
