+++ api/model/form/eyecolor_field.go
@@ -1,0 +1,9 @@
+package form
+
+// EyeColorField stores the persons eye color
+type EyeColorField string
+
+func (f EyeColorField) Valid() (bool, error) {
+	// Check if from enum of hair color
+	return true, nil
+}
