+++ api/model/form/month_year_field.go
@@ -1,0 +1,12 @@
+package form
+
+// MonthYearField is a subset of the DateField
+type MonthYearField struct {
+	Month     int64
+	Year      int64
+	Estimated bool
+}
+
+func (d MonthYearField) Valid() (bool, error) {
+	return true, nil
+}
