+++ api/model/form/url_field.go
@@ -1,0 +1,8 @@
+package form
+
+// URLField contains a properly formatted URL
+type URLField string
+
+func (e URLField) Valid() (bool, error) {
+	return true, nil
+}
